int换64位

+++ WhereToBuy-3/PersonCenterViewController.h
@@ -18,6 +18,6 @@
 
 @protocol PersonTopicAndFeedbackDelegate <NSObject>
 
-- (void)showTableViewType:(int)type;
+- (void)showTableViewType:(NSInteger)type;
 
 @end
