+++ WhereToBuy-3/WhereToBuy/Classes/SDWebImage/SDWebImageDecoder.m
@@ -25,7 +25,7 @@
     CGColorSpaceRef colorSpace = CGColorSpaceCreateDeviceRGB();
     CGBitmapInfo bitmapInfo = CGImageGetBitmapInfo(imageRef);
 
-    int infoMask = (bitmapInfo & kCGBitmapAlphaInfoMask);
+    NSInteger infoMask = (bitmapInfo & kCGBitmapAlphaInfoMask);
     BOOL anyNonAlpha = (infoMask == kCGImageAlphaNone ||
             infoMask == kCGImageAlphaNoneSkipFirst ||
             infoMask == kCGImageAlphaNoneSkipLast);
